[TC] Regenerated demo for tc_tc_compare_mode_sam_rh71_ek
MH3-68312

+++ apps/tc/tc_compare_mode/firmware/src/config/sam_rh71_ek/peripheral/tc/plib_tc0.c
@@ -112,7 +112,7 @@ void TC0_CH0_ComparePeriodSet (uint32_t period)
 /* Read the period value */
 uint32_t TC0_CH0_ComparePeriodGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[0].TC_RC;
+    return (uint32_t)TC0_REGS->TC_CHANNEL[0].TC_RC;
 }
 
 /* Set the compare A value */
@@ -188,7 +188,7 @@ void TC0_CH1_ComparePeriodSet (uint32_t period)
 /* Read the period value */
 uint32_t TC0_CH1_ComparePeriodGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[1].TC_RC;
+    return (uint32_t)TC0_REGS->TC_CHANNEL[1].TC_RC;
 }
 
 /* Set the compare A value */
